Remove old lib/index.js file
No longer needed with removal of babel
